Remove redundant serialization tests
