--- test/Microsoft.AspNetCore.JsonPatch.Test/TestObjectModels/JsonPropertyObject.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using Newtonsoft.Json;
-
-namespace Microsoft.AspNetCore.JsonPatch
-{
-    public class JsonPropertyObject
-    {
-        [JsonProperty("AnotherName")]
-        public string Name { get; set; }
-    }
-}
