--- test/Microsoft.AspNetCore.JsonPatch.Test/TestObjectModels/JsonPropertyWithAnotherNameObject.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-namespace Microsoft.AspNetCore.JsonPatch
-{
-    public class JsonPropertyWithAnotherNameObject
-    {
-        public string AnotherName { get; set; }
-    }
-}
