--- test/Microsoft.AspNetCore.JsonPatch.Test/TestObjectModels/SimpleObjectWithNullCheck.cs
@@ -1,30 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-
-namespace Microsoft.AspNetCore.JsonPatch
-{
-    public class SimpleObjectWithNullCheck
-    {
-        private string stringProperty;
-
-        public string StringProperty
-        {
-            get
-            {
-                return stringProperty;
-            }
-
-            set
-            {
-                if (value == null)
-                {
-                    throw new ArgumentNullException();
-                }
-
-                stringProperty = value;
-            }
-        }
-    }
-}
